Rename folder to modules from js
