+++ modules/localStorageTools.js
